DDA-69: Refactor grid tools
