+++ src/react/features/lenses/Lenses.js
@@ -2,8 +2,6 @@ import React from 'react'
 
 import { useStore } from '../../store'
 import { useFocusDiory } from '../diograph/hooks'
-import { useTools } from '../tools/useTools'
-import { useLensActions } from './useLensActions'
 
 import Fullscreen from '../../components/Fullscreen'
 
@@ -21,20 +19,12 @@ export const lenses = {
 
 const Lenses = () => {
   const { diory, diorys } = useFocusDiory()
-  const [{ active }] = useStore((state) => state.buttons)
   const [{ selectedLensId }] = useStore((state) => state.lenses)
   const { Lens } = lenses[selectedLensId]
-  const actions = useLensActions()
-  const { onClick } = useTools()
+
   return diory ? (
     <Fullscreen marginTop={48} zIndex={-1}>
-      <Lens
-        diory={diory}
-        diorys={diorys}
-        activeButton={active}
-        actions={actions}
-        toolOnClick={onClick}
-      />
+      <Lens diory={diory} diorys={diorys} />
     </Fullscreen>
   ) : null
 }
