+++ src/react/features/tools/Tools.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import CreateTool from './create/CreateTool'
+import UpdateTool from './update/UpdateTool'
+
+const Tools = () => (
+  <>
+    <CreateTool />
+    <UpdateTool />
+  </>
+)
+
+export default Tools
