+++ src/react/features/tools/move/useMoveTool.js
@@ -1,9 +1,11 @@
 import { useDispatch } from '../../../store'
 import { useButtons } from '../../buttons'
+import { useMoveToolIsActive } from './useMoveToolIsActive'
+
+import { setInactive } from '../../buttons/actions'
 import { updateDiory } from '../../diograph/actions'
 
 import { buttons } from './buttons'
-import { useMoveToolIsActive } from './useMoveToolIsActive'
 
 export const useMoveTool = () => {
   useButtons(buttons)
@@ -13,6 +15,7 @@ export const useMoveTool = () => {
   return (diory) => {
     if (isActive) {
       dispatch(updateDiory(diory))
+      dispatch(setInactive())
     }
   }
 }
