+++ src/react/features/tools/update/index.js
@@ -1,0 +1,1 @@
+export { useUpdateTool } from './useUpdateTool'
